try debug secrets
